* print true/false instead of 0/1 * support lifedrain and manadrain elements * remove "set" from several functions names for pyot monster

+++ OTMonsterConverter/Contracts.GenericMonster.cs
@@ -75,7 +75,7 @@ namespace OTMonsterConverter
     public interface ICustomSummon
     {
         string Name { get; set; }
-        ushort Rate { get; set; }
-        ushort Chance { get; set; }
+        uint Rate { get; set; }
+        double Chance { get; set; }
     }
 }
